Add or=00 to LS_COLORS
This avoids issues for ls in $HOME with symbolic links when scratch
is down.

By default symbolic links that point to an invalid location are
red on black, and valid ones are cyan.

This change makes the invalid ones cyan too, and avoids multiple
syscalls for plain "ls --color=auto".