Singularity/Apptainer: add custom bin directory to PATH
The unsquashfs binary in Gentoo is incompatible with Apptainer < 1.1
and Singularity 3.7/3.8. Force the use of /usr/sbin/unsquashfs
via a symbolic link in
/cvmfs/soft.computecanada.ca/custom/software/apptainer/bin
and adding that to PATH.

This didn't use to be necessary since Singularity/Apptainer rejected
the Gentoo unsquashfs because it did not support xattr. Now that it
does support xattr (for compatibility with 1.1) it runs into a
GLIBC_PRIVATE symbol issue with older versions on CentOS 7.