CCconfig: don't define TMPDIR, fix LOCAL_SCRATCH
TMPDIR set to SLURM_TMPDIR has issues with recursive sbatch invocations,
since Slurm complains about it not existing in inner jobs and CCconfig
will not adjust it then. As in 2017 the only reason for defining TMPDIR
was to avoid filling up / instantly using /tmp, but since then /tmp has
changed to a private tmpfs mount set up by a SPANK plugin, so we can just
leave TMPDIR untouched (as was the consensus in 2017 otherwise).

Slurm doesn't complain about LOCAL_SCRATCH but it still needs to be set
correctly in a recursive job and not be inherited as an invalid
directory name.