Get project quota for tamIA's /home and /scratch
The home and scratch file systems of the tamIA cluster are in the same
Lustre namespace. Quota are managed with project quota for Lustre.

The project id for a user home directory is the user's UID while the
project id for a user scratch directory is the user's UID plus an offset
of 20000000.